Remove key injection from CheckoutSessionRepository
All CheckoutSessionRepository methods now accept requestOptions as a
parameter. CheckoutController and SavedPaymentMethodRepository receive
ApiRequest.Options via DI from PaymentConfigurationModule.

Committed-By-Agent: claude
