+++ payments-core/src/main/java/com/stripe/android/payments/core/injection/PaymentConfigurationModule.kt
@@ -5,6 +5,7 @@ import androidx.annotation.RestrictTo
 import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.core.injection.STRIPE_ACCOUNT_ID
+import com.stripe.android.core.networking.ApiRequest
 import dagger.Module
 import dagger.Provides
 import javax.inject.Named
@@ -29,4 +30,12 @@ class PaymentConfigurationModule {
     fun provideStripeAccountId(
         paymentConfiguration: Provider<PaymentConfiguration>
     ): () -> String? = { paymentConfiguration.get().stripeAccountId }
+
+    @Provides
+    fun provideApiRequestOptions(
+        paymentConfiguration: Provider<PaymentConfiguration>
+    ): ApiRequest.Options = ApiRequest.Options(
+        apiKey = paymentConfiguration.get().publishableKey,
+        stripeAccount = paymentConfiguration.get().stripeAccountId,
+    )
 }
