+++ paymentsheet/src/main/java/com/stripe/android/checkout/injection/CheckoutModule.kt
@@ -2,11 +2,9 @@ package com.stripe.android.checkout.injection
 
 import android.app.Application
 import android.content.Context
-import com.stripe.android.PaymentConfiguration
 import com.stripe.android.common.di.DisplayDensity
 import com.stripe.android.core.injection.ENABLE_LOGGING
 import com.stripe.android.core.injection.StripeNetworkClientModule
-import com.stripe.android.core.networking.ApiRequest
 import com.stripe.android.payments.core.injection.PRODUCT_USAGE
 import com.stripe.android.payments.core.injection.PaymentConfigurationModule
 import com.stripe.android.paymentsheet.BuildConfig
@@ -15,7 +13,6 @@ import com.stripe.android.uicore.image.StripeImageLoader
 import dagger.Module
 import dagger.Provides
 import javax.inject.Named
-import javax.inject.Provider
 
 @Module(includes = [PaymentConfigurationModule::class, StripeNetworkClientModule::class])
 internal object CheckoutModule {
@@ -29,14 +26,6 @@ internal object CheckoutModule {
     @Provides
     @Named(PRODUCT_USAGE)
     fun provideProductUsageTokens(): Set<String> = setOf("Checkout")
-
-    @Provides
-    fun provideApiRequestOptions(
-        paymentConfiguration: Provider<PaymentConfiguration>
-    ): ApiRequest.Options = ApiRequest.Options(
-        apiKey = paymentConfiguration.get().publishableKey,
-        stripeAccount = paymentConfiguration.get().stripeAccountId,
-    )
 
     @Provides
     fun provideStripeImageLoader(context: Context): StripeImageLoader = DefaultStripeImageLoader(context)
