+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/intent/CheckoutSessionConfirmationInterceptor.kt
@@ -118,6 +118,7 @@ internal class CheckoutSessionConfirmationInterceptor @AssistedInject constructo
         return checkoutSessionRepository.confirm(
             id = integrationMetadata.id,
             params = params,
+            requestOptions = requestOptions,
         ).fold(
             onSuccess = ::handleConfirmResponse,
             onFailure = { error ->
